tests: Redesign tests to use sub-tests
we need to be able to reliably start and stop a TCP server and
TestMain() + go srv.Run() doesn't seem to work well b/c there's
no good way to do teardown (stop the server).

Hence group similar test scenarios and execute them as sub-tests
with a defered shutdown of the running TCP server. That should
make it possible for the next such test scenario to be able to
start the TCP server again.