CI: use osbuild and composer commits deployed in staging / production
Modify `schutzbot/deploy.sh` to try to use the actual commits of
osbuild and osbuild-composer deployed in staging / production for CI
runs on main (staging) and stable (production) branches. The same
applies for PRs to main and stable branches. For any other situation, use
the default values defined in `deploy.sh`.

Signed-off-by: Tomas Hozza <thozza@redhat.com>