LoadTesing: Implement a load test for image builder
The load testing is using the locust load test framework(https://locust.io).
The test is made to run inside the CI environment and against internal the stage
cluster.
The test is not made to be part of the PR test cycle instead it is made to run
on a schedule.

Since the test is only triggered on a schedule and not on a PR, the
commit was tested separately, you can have the result of the testing there:
https://gitlab.com/osbuild/ci/image-builder/-/jobs/1534065317