test: run api tests against stage
Rather than spin up a composer instance to test against, use
api.stage.openshift.com to make the tests as close to production
as possible.

This requires a COMPOSER_OFFLINE_TOKEN to be provisioned in the
gitlab CI runners. And the offline token must correspond to an
account that has been granted access in composer.