db: make org_id the primary tenant identifier
console.redhat.com is moving from account_number to org_id as the primary
tenant identifier.

The reason for this is that every account has an org_id, but not every account
has an account_number. The account_number is associated with an account
when a billing relationship is established (for instance when a subscription is
created for the account). console.redhat.com in general, and image-builder in
particular, wants to allow users access before they have a billing relationship.

This is the first step in that direction. Note that the account_number requirement
has not been dropped, and that will happen in a follow-up.