v1: limit the number of image requests in the spec itself
We have always supported only one image request per a compose. This is not
enforced on the openapi spec level though, just on the Go code level.
Let's add the limits also to the spec file to be more explicit.

The old check from server.go is removed because this is now enforced at the
API level. Also a test is modified because we now return a different message.

As passing more/no requests was never possible, this is not an API change.