Remove Test Edge EL8 test script
this is already covered by ostree-ng.sh, installers.sh and
ostree-simplified-installer.sh inside the osbuild-composer repository.
The only real difference here is that instead of building the images via
cli the script is using an API call for that. Everything else is
actually testing the image definition, not the Image Builder service.

Here is not the place for that, if more testing is needed for the edge
image definition is should be added into the osbuild-composer and/or
osbuild repositories.