distribution: load packages when loading a distribution
This is slightly hacky, distribution.ReadDistribution now also reads the list
of packages that belongs to the distribution and saves them in Architecture.

This allows us to read the (potentially huge) list of packages only once when
starting the server.

Also, the test repositories for the allow list tests now have empty
repositories, otherwise the new loader fail to load the distributions because
they don't have the package lists defined.