distribution: add an option to have a distribution without a package list
For Fedora, we don't want to have the package list living in the repository.

This has two reasons:
- This won't be in the frontend for now, so it's actually not needed.
  (we don't expect people to use the packages search functionality via the API
  much)
- Fedora has a ton of packages so this file would be super-huge.

Therefore, distributions got a new field `no_package_list`. When it's set to
true, image-builder will skip loading the package list for the particular
distro.

Architecture.FindPackages was modified not to crash when there's no package
list.

A new testing distro was added in order to test this scenario.