distributions: add fedora
Let's add all versions that currently exist - 35-38. They don't have a package
list included, so package search won't work but that's fine, since it will be
only in the API. Also, it's marked as restricted, so before anyone can use it,
we will need to add an entry to the allow-list firstly.