Enable testing in pr_check.sh
We can now run IQE plugin tests on PRs in ephemeral environment. This
script is filled with defaults as that serves our needs just fine. It
builds the container image with the changes in the PR, provisions an
ephemeral namespace, deployes image-builder, runs the IQE
image-builder-backend test suite and reports back on the PR.