templates: add FloorPlan
FloorPlan for data extraction based on sample [1] and Compliance's
FloorPlan [2].

The query already runs weekly and is defined in app-interface.  With
this FloorPlan, we will configure the Floorist operator to store the
data in an S3 bucket in the Parquet format.

Include a variable to control the prefix for the storage location of the
Floorist data dump so that we can differentiate between stage and prod.

[1] https://github.com/RedHatInsights/floorist-operator/blob/main/config/samples/metrics_v1alpha1_floorplan.yaml
[2] https://github.com/RedHatInsights/compliance-backend/blob/822e2d9823d88a2ac01b2184c22e3a77827d411c/clowdapp.yaml#L272-L298