internal/v1: correctly deal with offsets in GetPackages
If the offset was 0, which is the default, and if no packages were
found, it panicked due to an out of bounds index.

The reasoning by subtracting one was that even if you chose too large of
an offset, it would still give you back something. Simply cap the offset
in case it's larger than the amount of packages. The data returned will
be empty but that's ok.

Cap the offset in the "last" link at 0 if no results were found.

The offset in the "first" link should always be 0.