internal/v1: add ocm metrics middleware
Add middleware to handle and track all our
routes and their status codes and not just
the `/compose` endpoint. This results in
a new prometheus metric that contains info
on the status code, path & request method.

A helper function has been added to clean
the dynamic parameters in the path routes
to reduce metric cardinality.