v1: add echo/middleware package and deps
Logging without correlation information is not the best practice for services under heavy load since there is nothing you can do to find logs which belong to each other. The standard practice is to add some sort of correlation id or request id. On our platform, that is the X-Rh-Edge-Request-Id which is set by 3scale and can be propagated through services. This patch introduces new dependencies.