internal/v1: accept source to share an Azure image with
This commit follows 019d2c8 which added support for sharing AWS images
with sources.

The /sources/{sourceId}/account_identity endpoint, which was used to
resolve the account id for AWS images shared with sources, is
deprecated.

This commit removes the deprecated endpoint, replacing it with the
new upload_info endpoint which is used for both AWS and Azure images.