api: embed the parent compose ID in clone response
The frontend uses RTKQ to perform its requests to the IB's backend, and
we're using programmatically generated requests to fetch the data over.
RTKQ allows us to enhance these automatically generated endpoints and
add a few behavioral customizations like cache tagging and cache
invalidation.

We are planning on using cache tagging and invalidation to enhance the
user experience by allowing clones to get refreshed only in certain
circumstances.

However, in the current state of the API we do not have access to the
parent composeID of a clone at the CloneResponseItem level, meaning that
our tag invalidation will invalidate too broadly.

Adding the parent composeID to the clone data that is returned will
allow us to only refetch the clone data associated to that given
composeID and only regenerate parts of the screen associated with that
data.