bib: add cross arch building
This commit adds experimental cross arch building.

I.e. you `bootc-image-builder` on a mac and target an `amd64` machine.

The usage would be:
```
sudo podman run ... \
  --type qcow2 \
  --target-arch amd64 \
     quay.io/centos-bootc/fedora-bootc:eln
```
and this is predicated on the idea that the target container is
available for both host and target architecture. When that is the case
the container for the host architecture is used as a buildroot, the
container for the target architecture is the target tree. Note that some
stages chroot into the target tree so `qemu-user` is essential for this
to work so that target achitecture code can be run transparently (because
only very small amounts of code need to run emulated this is pretty fast
still).

It will need qemu-user{,-static} installed on the host. Sadly we
cannot just install qemu-user into the container because the binfmt_misc
system is global and not namespaced. So while adding the binfmt_misc
handler from inside the container would work with enough privs it
would alter the global machine state and we should not do it.