Rename `iso` -> `anaconda-iso`
I would like to potentially introduce a generic "live iso" that
generates *exactly the input container* just as a Live ISO, very
similar to how Fedora CoreOS (and one of its parents, the original
Container Linux) do it.

This has several use cases, such as always running from a PXE
boot.  I also think this is the long term architecture for an ISO,
as opposed to the current Anaconda.

This "direct" Live ISO is also the same as how e.g. Fedora Workstation
does it, where the installer is just an app in the system.

For compatibility we continue to honor the `iso` as an alias.