bib: add --roofs to be able to select the root fs type
Some container images don't have a default filesystem type. In this
case, a disk image build will fail. I'm a strong believer that we need
a one-line command to build a disk image. Thus, this commit adds
--rootfs which sets the root filesystem type if the container image
doesn't have one (or overrides it if it has one). By using --rootfs,
users will be able to run just one command to build an image once
again.

In the future, we will support custom partitioning via blueprints (or
different format). This means that we will need to figure out what takes
precedence. However, I suggest to cross that bridge when we get there.