internal: include the http status on GetComposeStatus errors
This commit tweaks the error returned from GetComposeStatus()
when an unexpected http return code appears so that it includes
the acual http status code. This makes it easier for us to
debug what is going on on the other side when this error happens.

Note that if there are concerns about leaking this to users we
could also just log it but I don't see harm in showing it to
the users (as we are the users :)