internal: make `cClient.Compose()` retryable
Make the `cClient.Compose()` function retrying too. This requires
that we create a new `postBodyWithRetry()` helper because apparently
we get `200 OK` results that are actually errors so we cannot just
naively use the same helper and allow both 200,201. This might be
just the tests doing this and the real osbuild-composer is fine
but this commit is conversative.