client: implement retries on the client request level
This is a quick draft of the idea of Florian (thanks!) about moving
the retry handlign into the client.go:request() helper.

What is nice about it is that it's very local. It changes the
responsibilities of the library though, because before client.go
did (for the most part) not care about the http status code and
all error are handled on the upper layers. By adding retry the
library layer now needs to make policy decisions based on the
status code. If we go this way we should move more of the
error handling into the library I think.

The other open question in this commit is how to handle retries
if reading the body fails. Here we need to also think about
the semantic of POST and make sure that the other side will
not create multiple jobs if we sent the POST correctly but
did not get the full reply.

And this needs *tests* - it's just a quick draft to get an
idea of the drawbacks/advantages compared to PR#1268