bib: enable Anaconda Users module for ISO
We generally only enable the Users module in ISOs when a user is
configured because we want the installation to be fully automatic /
unattended and, in the general case, Anaconda will block and wait for
user input if no users are configured.  However, since we always add
`rootpw --lock` to our kickstart file, this satisfies Anaconda's user
configuration rule and the installation proceeds without blocking.

Enabling the Users module unconditionally allows users to include user
configuration when they inject their own kickstart files.