image: correctly set root minsize and type
With the changes in 5f704173af3741eb5385341dd0a7f499e3deeced, the
minimum size for the root filesystem, calculated as double the size of
the base container, was not being considered at all—the Filesystems
property of the ManifestConfig was being ignored.
Also, the rootfs type from the container was not being applied to the
root filesystem in the base partition table.

Replace the Filesystems property with a simple RootfsMinsize to clarify
its purpose and use it to update or add to the user-defined
customizations.

Add a new function, setRootfsType(), which sets the root filesystem type
based on the rootfs option set either by the container config or the
user option.
The function replicates the checks that were moved to
TestBasePartitionTablesHaveRoot().  The checks in the test code will
catch any modifications to the static partition tables defined in code.
Running the same checks while modifying the partition table will also
catch any issues caused by the partition table being modified before the
rootfs type is applied.