bib: set the RootFSType for cross-arch builds
Before this change, the RootFSType validation would set the option to an
empty string when an unsupported type was used for cross-arch builds.
This left the type for the root filesystem unchanged from the default
when generating the partition table for the manifest.
Now that the RootFSType is required, let's be explicit about what we're
doing and set it to the only (currently) supported type for cross-arch
builds: ext4.