build: bump oapi-codegen to 2.4.1 and fix tools bin
This change unpins oapi code generator via the sed hack and changes the
installation directory to tools/bin so we all use the same version of
the CLI when calling the script. It also installs a specific version
which is independent to `go.mod` so we will have to bump it from time to
time. This should bring better consisntency in regenerating sources.