internal: move `cntdnf` code back into `container`
Ondrej suggested to remove the `cntdnf` package again and move
the code back into container. My original thinking was to have
a separate package because it a "container" should not need to
have knowledge about dnf and we could have a `cnfapt` later but
then YAGNI and we can always split it out later.