v1: expose groups in user customization
When adding a user, immediately add it to the specified groups.

This patch technically includes a behavioural change (the wheel group
will no longer be added by default). But the main consumers of this
api (the edge-api) already add the wheel group each time, and the api
doesn't have `additionalProperties` disabled for the users
customization.