image-builder-maintenance: initial version (HMS-4245)
image-builder-maintenance for now runs vacuum
and cleans all composes older than 2 years (by default).
To remove customer related data according to the
retention policy.

As this is not really urgent, it's scheduled once
a week on Tuesdays to avoid problems on weekends.

Also introduces a github action test if
the container can still be built.