bib: load repo keys using external paths
Since d192f85ae9e5b6e5665b8ea1ae39677b24a9352a, we switched to using dnf
(osbuild-depsolve-dnf) from BIB itself, but using the base (bootc)
container root to read configurations, certificates, keys, etc.
Unfortunately, what we didn't change, is the way in which we load the
repository certs after the depsolve.

The extractTLSKeys() function currently takes a Container instance which
implements ReadFile() by `cat`ing a file from inside the running
container.  However, the paths we get from the depsolve are relative to
the BIB worktree.  Therefore, we can simply read the files directly.

Add a SimpleFileReader that implements the fileReader interface to
simply read a file from a path.  We could simplify this to not require
a reader interface at all, but let's keep the implementation in case we
ever need to switch back to the older way of depsolving from inside the
container.

Resolves COMPOSER-2408