clients/content_sources: get repos by both repo URL and repo ID
This offloads most of the logic to the content sources client. This
paves the way for supporting repositories with an upload origin, which
only have a repository ID available.

Furthermore this will allow us to treat the external origin repositories
in the same way, based on ID, whilst only using the URLs for the red hat
repositories.