refactor: introduce common logging
This is the first patch in the series that will bring common logging
across osbuild projects: image-builder, composer, composer-worker.

It builds on top of the new log/slog library which is part of Go from
1.21. By extracting handlers and common initialization and configuration
code into a common library, this makes sure logging is consistent across
all projects. To make the trasition smooth, a logrus proxy is introduced
which proxies all logrus logging calls to slog so changes can be
gradually performed.

It utilizes a middleware from strc tracing package, but tracing is
currently not enabled.