main: add `--bootc-defaultfs` option
Not all bootc containers declare a default filesystem, most noatbly
fedora does not. So this commit adds `--bootc-defaultfs` to set
it.

It *might* be interesting in the future to consider making this
more broad, i.e. to allow `--defaultfs` for package based distros
too (with the generic distro we would set it via
`DistroYAMl.DefaultFStype`). But definitely followup material
(and requires more discussion if we want it).

Closes: https://github.com/osbuild/image-builder-cli/issues/323