doc: initial repository blurb
After a chat in our community Matrix channel I realized that we have not
documented the various ways of managing repositories in `image-builder`.
Let's start that off with a small initial page that I will build on in
the following commits.

This relies on the implementation of the new `force-repo-dir` behavior
as that's the behavior documented [1].

[1]: https://github.com/osbuild/image-builder-cli/pull/408

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>