main: rename `data-dir` to `force-data-dir`
Let's name arguments that fully override built-in definitions with the
`force-` prefix. This makes room for an `extra-data-dir` in the future
that allows for drop-in data.

The same will be applied to passing other arguments in the future.

We use a normalization to rename the old argument to the new one so we
can keep accessing it under a single name and a deprecation to hide the
previous argument name but still accept it on the command.

Cobra will emit a deprecation warning for us.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>