test: add test that can run subscribed container test
We need a way to test if image-builder can work with subscribed
content when run via a container. This requires a pair of secrets
SUBSCRIPTION_{ORG,ACTIVATION_KEY} and a host environment that
has `subscription-manager` installed.

Note that just running `image-builder manifest` is not enough,
even when this works the librepo/curl download can still fail
if osbuild does not find the right secrets.

Thanks to lzap for reporting this.