bib: drop bibManifestFromCobraFor into bibManifestFromCobra
It used to make sense to have these different functions but
now that we have all bootc image types in the images library
the creation of the manifests is uniform so we can just do
it all in the same function.

This has the added benefit that it uncovered a bug that we
do not honor --rpmmd :( This is fixed with this commit.