bib: fix anaconda-iso mTLS key extraction
There was a (subtle) bug in the ibcli version of the mTLS
key exaction. It was happening too late, i.e. when the
bootc container was already unmounted.

This commit moves the extraction into the `Depsolve`
function which is run while the container is mounted
which means we can extract the mTLS config.

Note that this was not discovered earlier because we
lack and end-to-end test for RHEL based bootc images :(