github: move rhel test into its own pull_request_target action
The rhel_test.py was not working before because I (wrongly)
assumed that during a merge queue merge the branch would
have access to the repository secrets (which IMNSHO would
make sense given that the branch was approved and is going
into main). But that is not the case so this commit moves
to the same pattern we use in bib, i.e. we only allow this
to run for people with write permissions to the repo.

This means that the inital workflow for external contributors
will always fail but when a contributor clicks on retry it
will work.