cmd: remove bibupload as it was never exposed to the public
The bibupload command was originally part of bootc-image-builder
and when we moved the go code it moved here. But after talking
to Ondrej we decided it can be removed because it was never
part of the bootc-image-builder container. It is also problematic
as we don't have enough information from the image alone how
to upload it (i.e. what boot configuration etc). So lets just
remove it, *if* (and that is a big if) we ever need it again
we can alwas resurrect it from the git history.