Containerfile{.bib}: move build logic into Makefile
Now that we have two places that need to run the same build
comamnd just consolidate things into the Makefile under
`make build`.

Note that this cannot move into the makefile as for bib we want
to build *with* cgo to have the chmod/chown logic work as those
need cgo/libnss (in the general case).