many: `force-repo-dir` instead of `force-data-dir`
Previously I deprecated `data-dir` and renamed it to `force-data-dir` to
have consistent naming for options that overwrite internal state instead
of appending to it (arguments that start with `force-` are overwrites,
arguments that start with `extra-` are merges).

With `force-data-dir` we require a `repositories` sub directory. After
some discussion we came to the idea to instead of enforcing a directory
layout on the user we want to be able to point directly at a directory
of repositories.

This commit deprecates `force-data-dir` and introduces `force-repo-dir`.
Both write to the same internal variable.

We also allow repository files to live *either* in a `repositories`
subdirectory to keep backwards compatibility for any of the arguments or
to live directly at the top level of the passed directory.

We show a deprecation warning if a `repositories` sub directory exists
and inform users that the `repositories` subdirectory will not be wanted
in the future and how to resolve the emitted warning.

This option will be used together with the future introduction of
`force-defs-dir` to handle definitions (and its `extra-defs-dir` counter
part) unifying all options into a consistent naming scheme.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>