test: drop `test_progress_term_autoselect` as it moved to ibcli
This test is no longer working, it seems its a result of the
merge of
osbuild/image-builder-cli#316

Drop it for now so that we can move forward with the merge.
There is a similar test in
https://github.com/osbuild/image-builder-cli/pull/316/files#diff-57c49abc8a31b46856c2eecb2edfeff6072cd8d0edf769a67b6d4fdaabd8321aR100
that does not cover "auto" but any fixes/regression should be
done in the ibcli repo.

See also https://github.com/osbuild/bootc-image-builder/pull/1157