main: print pretty json manifest
This commit makes us json pretty the osbuild manifests.
This means that when `image-builder manifest` or
`image-builder build --with-manifest` is used we get a
nice representation of the manifest (and it has no
consequences otherwise except for a few more bytes in
memory for the extra whitespace).

Given that people who ask for a manifest often want to
inspect it seems nice to make it pretty by default. OTOH
there is jq - so if its considered fluff that is fine too :)